lang:rust: update

+++ lang/rust/ffi1.rs
@@ -1,6 +1,7 @@
 
 extern {
-  fn write(fd: i32, data: *const u8, len: usize) -> i32;
+  // write(2): ssize_t write(int fd, const void *buf, size_t count);
+  fn write(fd: i32, data: *const u8, len: usize) -> usize;
 }
 
 fn main() {
